Changed directory of check_preference() function, removed utils from test directory
